// src/data/VideoData.d.ts
export declare const VideoDataSingle: {
    id: number;
    title: string;
    source: string;
    tags: {
        category: {
            id: number;
            name: string;
            slug: string;
        }[];
    };
};
export declare const VideoData: {
    id: number;
    title: string;
    source: string;
    tags: {
        category: {
            id: number;
            name: string;
            slug: string;
        }[];
    };
}[];
//# sourceMappingURL=VideoData.d.ts.map